# openapi.yaml
openapi: '3.0.2'
info:
  title: "Annotation tool"
  description: "The YAML API specification of the Annotation Tool, developed as part of the Software Project course."
  version: 1.0.0
servers:
  - url: https://localhost:8080
    description: Local server
...
paths:
...
  /submissions/{id}:
    get:
      summary: Get a submission
      description: Retrieve a submission from the database
      tags:
        - Submissions
      parameters:
        - in: path
          name: "id"
          description: "ID of the submission to be retrieved"
          required: true
          schema:
            type: string
            example: "student@tudelft.nl"
      responses:
        '200':
          description: "Successfully retrieved submission"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDTO"
        '404':
          description: "Could not find the specified submission in the database"
    put:
      summary: Update submission
      description: Update a submission in the database
      tags:
        - Submissions
      parameters:
        - in: path
          name: "id"
          description: "ID of the submission to be retrieved"
          required: true
          schema:
            type: string
            example: "student@tudelft.nl"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionDTO"
      responses:
        '200':
          description: "Successfully updated submission"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDTO"
        '400':
          description: "Could not update submission"
    delete:
      summary: Delete submission
      description: Delete a submission from the database
      tags:
        - Submissions
      parameters:
        - in: path
          name: "id"
          description: "ID of the submission to be deleted"
          required: true
          schema:
            type: string
            example: "student@tudelft.nl"
      responses:
        '200':
          description: "Successfully deleted submission"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionDB"
        '404':
          description: "Could not find submission"
  /submissions/export/{id}:
    get:
      summary: Download a submission
      description: Download a submission with the specified id
      tags:
        - Submissions
        - Supervisor
      parameters:
        - in: path
          name: "id"
          description: "ID of the submission to be retrieved"
          required: true
          schema:
            type: string
            example: "student@tudelft.nl"
      responses:
        '200':
          description: "Successfully retrieved submission"
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: "Could not find the specified submission in the database"
        '500':
          description: "Could not convert submission to byte array"
  /submissions/coordinator/{id}:
    get:
      summary: Retrieve coordinator's submissions
      description: Retrieve all submissions associated to the coordinator with the specified id
      tags:
        - Submissions
        - Supervisor
      parameters:
        - in: path
          name: "id"
          description: "ID of the coordinator"
          required: true
          schema:
            type: string
            example: "supervisor@tudelft.nl"
      responses:
        '200':
          description: "Successfully retrieved submissions"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionDTO"
        '500':
          description: "Could not retrieve all submissions"
  /submissions/search/{text}/{coordinator}:
    get:
      summary: Search submissions
      description: Search for all submissions that contain a piece of text and are assigned to some coordinator
      tags:
        - Submissions
        - Supervisor
      parameters:
        - in: path
          name: "text"
          description: "Text that the submission's email must contain"
          required: true
          schema:
            type: string
            example: "stud"
        - in: path
          name: "coordinator"
          description: "ID of the coordinator who is responsible for the submission"
          required: true
          schema:
            type: string
            example: "supervisor@tudelft.nl"
      responses:
        '200':
          description: "Successfully retrieved submissions"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionDTO"
...
components:
  schemas:
    User:
      type: object
      properties:
        email:
          type: string
          example: "supervisor@tudelft.nl"
        name:
          type: string
          example: "Supervisor One"
        password:
          type: string
          example: "superPass"
        role:
          type: string
          example: "supervisor"
        correspondingSubmissions:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionDB"
    SubmissionDB:
      type: object
      properties:
        email:
          type: string
          example: "student@tudelft.nl"
        file_submission:
          type: string
          format: binary
        group_name:
          type: string
          example: "test group"
        assignedCoordinators:
          type: array
          items:
            $ref: "#/components/schemas/User"
        file_name:
          type: string
          example: "submission.pdf"
        last_submitted:
          type: string
          example: "Tue, 10 Jan 2007 16:42:00 GMT"
        last_edited:
          type: string
          example: "Wed, 10 Mar 2006 08:12:22 GMT"
        is_submitted:
          type: boolean
          example: true
        is_locked:
          type: boolean
          example: false
...
    SubmissionDTO:
      type: object
      properties:
        id:
          type: string
          example: "student@tudelft.nl"
        file_submission:
          type: string
          format: byte
        group_name:
          type: string
          example: "test group"
        assignedCoordinators:
          type: array
          items:
            $ref: "#/components/schemas/User"
        file_name:
          type: string
          example: "submission.pdf"
        last_submitted:
          type: string
          example: "Tue, 10 Jan 2007 16:42:00 GMT"
        last_edited:
          type: string
          example: "Wed, 10 Mar 2006 08:12:22 GMT"
        is_submitted:
          type: boolean
          example: true
        is_locked:
          type: boolean
          example: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []